ai-data/requirements.txt
streamlit
pandas
langchain
langchain-experimental
langchain-google-genai
python-dotenv
tabulate